re #1171 korzystanie z opisu swiata - generowanie kolejnych plansz, wip

+++ src/GUI/CHud.cpp
@@ -43,7 +43,7 @@ CHud::~CHud()
 {
 }
 
-void CHud::Init(unsigned int playerNumber)
+void CHud::Init(size_t playerNumber)
 {
 	std::string prefix = std::string("pl") + StringUtils::ToString(playerNumber) + "-";
